alterações do crush

+++ app/src/main/java/br/senai/rn/agenda/activities/ListaAlunosActivity.java
@@ -5,6 +5,8 @@ import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v7.app.AppCompatActivity;
+import android.view.ContextMenu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
@@ -19,7 +21,7 @@ public class ListaAlunosActivity extends AppCompatActivity {
     private ListView listaAlunos;
     private FloatingActionButton botaoAdicionar;
     private ArrayAdapter<Aluno> adapter;
-    private AlunoDAO dao = new AlunoDAO();
+    private AlunoDAO dao;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -33,13 +35,20 @@ public class ListaAlunosActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
-        adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1, dao.obterTodos());
-        listaAlunos.setAdapter(adapter);
+        adapter.clear();
+        adapter.addAll(dao.obterTodos());
     }
 
     private void inicializarComponentes() {
+        dao = new AlunoDAO();
+
         listaAlunos = findViewById(R.id.principal_activity_main_lista_de_alunos);
         botaoAdicionar = findViewById(R.id.principal_activity_main_fab_novo_aluno);
+
+        adapter = new ArrayAdapter<>(this, android.R.layout.simple_list_item_1);
+        listaAlunos.setAdapter(adapter);
+
+        registerForContextMenu(listaAlunos);
 
         dao.salvar(new Aluno("Miguel", "998998876", "adsfkafdja@gmail.com"));
         dao.salvar(new Aluno("Marcos", "3453535", "sla@gmail.com"));
@@ -64,15 +73,29 @@ public class ListaAlunosActivity extends AppCompatActivity {
                 startActivity(intent);
             }
         });
-        listaAlunos.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
-            @Override
-            public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
-                Aluno alunoSelecionado = (Aluno) parent.getItemAtPosition(position);
-                dao.remover(alunoSelecionado);
-                adapter.remove(alunoSelecionado);
-                return true;
-            }
-        });
     }
 
+    @Override
+    public void onCreateContextMenu(ContextMenu menu, View v, ContextMenu.ContextMenuInfo menuInfo) {
+        super.onCreateContextMenu(menu, v, menuInfo);
+        getMenuInflater().inflate(R.menu.activity_lista_aluno_menu, menu);
+//        menu.add("Remover");
+    }
+
+    @Override
+    public boolean onContextItemSelected(MenuItem item) {
+        int menuSelecionado = item.getItemId();
+
+        if (menuSelecionado == R.id.activity_lista_aluno_menu_remover) {
+            AdapterView.AdapterContextMenuInfo menuInfo
+                    = (AdapterView.AdapterContextMenuInfo) item.getMenuInfo();
+            Aluno alunoSelecionado = adapter.getItem(menuInfo.position);
+            dao.remover(alunoSelecionado);
+            adapter.remove(alunoSelecionado);
+        }
+
+
+        return super.onContextItemSelected(item);
+
+    }
 }
